refactor(api): split backtest monitoring into focused services
- Extract analytics, query, export, and shared utilities into dedicated services
- Split signal, trade, optimization, live-replay, and paper-trading monitoring
- Add comprehensive spec coverage for each new service
- Wire new services through admin module
- Extract shared CSV serialization into admin/utils/csv.util with tests
- Centralize MAX_EXPORT_LIMIT in admin/constants and reuse across services
- Simplify export and live-trade-monitoring services via shared CSV helper
- Fix algorithmId filter to use direct FK column instead of relation navigation

Closes #333
