+++ apps/api/src/admin/constants.ts
@@ -1,0 +1,1 @@
+export const MAX_EXPORT_LIMIT = 10000;
